默认关闭线程组配置导入功能 线程组配置目前只支持默认线程组和少量的扩展线程组，为了防止大家滥用，默认关闭这个功能，如果要开启通过参数JMETER_THREADGROUP_SET_KEY为true开启。 权重功能只针对默认线程组，不支持扩展线程组（主要是扩展线程组千奇百怪，对应的配置参数也多，权重调整不好把控，如果需要可自定义开发：在LocalDistributedRunner.java中调入扩展的线程组类进行参数权重调配）

+++ src/main/java/io/renren/modules/test/service/impl/StressTestFileServiceImpl.java
@@ -219,8 +219,8 @@ public class StressTestFileServiceImpl implements StressTestFileService {
         // 肯定存在已有的用例信息
         stressTestDao.update(stressCase);
         stressTestUtils.saveFile(multipartFile, filePath);
-        // 对jmx脚本将线程组配置信息入库
-        if(filePath.substring(filePath.length()-3).equals("jmx")){
+        // 对jmx脚本将线程组配置信息入库(默认不入库)
+        if(stressTestUtils.isGetThreadGroup() && filePath.substring(filePath.length()-3).equals("jmx")){
         	try {
         		//入库前清理已有配置项
         		testStressThreadSetDao.deleteByFileId(stressTestFile.getFileId());
